+++ src/test/kotlin/com/zepben/auth/client/ZepbenTokenFetcherTest.kt
@@ -170,20 +170,13 @@ internal class ZepbenTokenFetcherTest {
     }
 
     @Test
-    fun testFetchTokenSuccessful() {
-        doReturn(StatusCode.OK.code).`when`(response).statusCode()
-        doReturn("{\"access_token\":\"$TOKEN\", \"token_type\":\"Bearer\"}").`when`(response).body()
+    fun testFetchAuth0TokenSuccessful() {
+        testSuccessfulFetch(AuthMethod.AUTH0, "audience=test_audience")
+    }
 
-        val tokenFetcher = ZepbenTokenFetcher(
-            audience = "test_audience",
-            tokenEndpoint = "https://testissuer.com.au",
-            authMethod = AuthMethod.OAUTH,
-            client = client,
-        )
-        verify(client, never()).send(any(), any<HttpResponse.BodyHandler<String>>())
-        val token = tokenFetcher.fetchToken()
-        verify(client).send(any(), any<HttpResponse.BodyHandler<String>>())
-        assertThat(token, equalTo("Bearer $TOKEN"))
+    @Test
+    fun testFetchEntraTokenSuccessful() {
+        testSuccessfulFetch(AuthMethod.ENTRAID, "scope=test_audience/.default")
     }
 
     @Test
@@ -279,27 +272,38 @@ internal class ZepbenTokenFetcherTest {
     }
 
     @Test
-    fun testFetchTokenSuccessfulUsingRefresh() {
+    fun testFetchAuth0TokenSuccessfulUsingRefresh() {
+        testSuccessfulFetch(AuthMethod.AUTH0, "audience=test_audience", "refresh_token")
+    }
+
+    @Test
+    fun testFetchEntraTokenSuccessfulUsingRefresh() {
+        testSuccessfulFetch(AuthMethod.ENTRAID, "scope=test_audience/.default", "refresh_token")
+    }
+
+    private fun testSuccessfulFetch(authMethod: AuthMethod, expectedScope: String, refreshToken: String? = null) {
         doReturn(StatusCode.OK.code).`when`(response).statusCode()
         doReturn(
-            "{\"access_token\":\"$TOKEN\", \"refresh_token\": \"test_refresh_token\", \"token_type\":\"Bearer\"}"
+            "{\"access_token\":\"$TOKEN\", ${refreshToken?.let { "\"refresh_token\": \"$it\"," } ?: ""} \"token_type\":\"Bearer\"}"
         ).`when`(response).body()
 
         mockStatic(BodyPublishers::class.java, CALLS_REAL_METHODS).use { bodyPublishers ->
             val tokenFetcher = ZepbenTokenFetcher(
                 audience = "test_audience",
                 tokenEndpoint = "https://testissuer.com.au",
-                authMethod = AuthMethod.OAUTH,
+                authMethod = authMethod,
                 client = client,
-                refreshToken = "test_refresh_token"
+                refreshToken = refreshToken
             )
             verify(client, never()).send(any(), any<HttpResponse.BodyHandler<String>>())
             val token = tokenFetcher.fetchToken()
             verify(client).send(any(), any<HttpResponse.BodyHandler<String>>())
-            bodyPublishers.verify { BodyPublishers.ofString(matches("audience=test_audience&refresh_token=test_refresh_token")) }
-            assertThat(token, equalTo("Bearer $TOKEN"))
-        }
+            bodyPublishers.verify {
+                BodyPublishers.ofString(matches("${expectedScope}${refreshToken?.let { "&refresh_token=$it" } ?: ""}"))
+                assertThat(token, equalTo("Bearer $TOKEN"))
+            }
 
+        }
     }
 
     @Test